pyproject.toml
[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "cc"
module-root = ""

[project]
name = "confident-confidant"
version = "2.0.0"
description = "Use spooky math to transform audio notes into transcripts and summaries"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "hjson",
  "litellm",
  "openai>=2.0.0",
  "thds-core",
  "thds-humenc",
  "thds-mops",
]

[dependency-groups]
dev = ["ruff", "mypy", "mypy-extensions", "pytest"]

[tool.ruff]
line-length = 105

[project.scripts]
# cc is already a program on my PATH
coco = "cc.__main__:main"
coco-summarize = "cc.__main__:main_summarize"
transcribe = "cc.transcribe.__main__:cli"
transcribe-diarize = "cc.transcribe.diarize.__main__:cli"
transcribe-label = "cc.transcribe.diarize.label:cli"

[tool.mypy]
mypy_path = "."
namespace_packages = true
explicit_package_bases = true
exclude = '''(?x)(
    build/
    | dist/
    | \S+\.egg-info/
)'''
check_untyped_defs = true
disallow_untyped_calls = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = ["hjson"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
markers = ["e2e: end-to-end tests (hit real APIs, require ffmpeg)"]
addopts = "-m 'not e2e'"
